This is a potential underflow. Fixing.

+++ src/EGaussian.cpp
@@ -114,7 +114,7 @@ EGaussian::~EGaussian() {
 
 void EGaussian::canceling(const uint32_t sublevel) {
     uint32_t a = 0;
-    for (int i = clauses_toclear.size() - 1; i >= 0 && clauses_toclear[i].second > sublevel; i--) {
+    for (int i = (int)clauses_toclear.size() - 1; i >= 0 && clauses_toclear[i].second > sublevel; i--) {
         solver->cl_alloc.clauseFree(clauses_toclear[i].first);
         a++;
     }
